Code reorganization: make the code modular
 - Add BruteForce.cpp : for brute Force algorithms
 - Add utility.cpp & utility.h and parameter.h files
